requirements.txt
Django>=2.2
python-social-auth
html2markdown
Markdown==3.1.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
psycopg2-binary==2.8.4
python-decouple==3.1
dj-database-url==0.5.0
django-widget-tweaks==1.4.5
requests
beautifulsoup4
isodate
